PyDict: implement max_size(), fix insert() and erase()

+++ src/python/tests/PyDict_test.cc
@@ -82,18 +82,18 @@ public:
     CPPUNIT_ASSERT_EQUAL((Py_ssize_t) 1, dict.object()->ob_refcnt);
 
     CPPUNIT_ASSERT(dict.empty());
-    CPPUNIT_ASSERT_EQUAL(0, dict.size());
+    CPPUNIT_ASSERT_EQUAL((size_t) 0, dict.size());
 
     dict["test"];
-    CPPUNIT_ASSERT_EQUAL(1, dict.size());
+    CPPUNIT_ASSERT_EQUAL((size_t) 1, dict.size());
     CPPUNIT_ASSERT(dict["test"] == Py_None);
 
     dict["test"] = 2;
-    CPPUNIT_ASSERT_EQUAL(1, dict.size());
+    CPPUNIT_ASSERT_EQUAL((size_t) 1, dict.size());
     CPPUNIT_ASSERT(dict["test"] == 2);
 
     dict["test"] = "string";
-    CPPUNIT_ASSERT_EQUAL(1, dict.size());
+    CPPUNIT_ASSERT_EQUAL((size_t) 1, dict.size());
     CPPUNIT_ASSERT(dict["test"] == "string");
     CPPUNIT_ASSERT_EQUAL((Py_ssize_t) 1, dict.object()->ob_refcnt);
